Scaffolding will become a own library
The scaffolding library currently has a lot of issues, so we will make
a own library out of this.
Code-vise this has now connection to the Bancha project, so a split
into two project is very easy. If you are currently using the scaffold
library please contact me (mail@rolandschuetz.at), the new lib will be
online soon.